Move GPU-specific kernel sources into src/core/gpu_kernel/
The CPU kernels already live under src/core/cpu_kernel/. Mirror that layout
on the GPU side so the next round of GPU work (per-arch knobs + autotuning,
issue #115) has a clear home and is not interleaved with the host-side
tile/profile orchestration.

Files moved (git rename, no behavior change):
  kernels.h, kernels.cu                 -- main kernel entry + dispatch
  kernels_compute.h, kernels_smem.h     -- per-profile compute strategies + smem
  kernel_gpu_utils.h, kernel_gpu_utils.cu -- launch config helpers
  qt_kernels.h, qt_kernels.cu           -- cuFFT sliding-dot-product kernels

Files staying at src/core/ (host-side or shared):
  kernel_common.h/.cpp -- shared with CPU kernels
  qt_helper.h/.cpp     -- host-side cuFFT context manager (has CPU fallback)
  tile.h/.cpp          -- tile orchestration

src/core/gpu_kernel/CMakeLists.txt now owns the qt_kernels, gpu_utils, and
gpu_kernels library targets and is only added when CMAKE_CUDA_COMPILER is
present, mirroring how cpu_kernel/ is structured.

Verified that build_cuda/ builds cleanly with -DFORCE_CUDA=1 and that the
1NN_INDEX self-join on a 2000-element series exactly matches
distance_matrix_np reference output (same vectors and indexes).
