+++ src/core/gpu_kernel/kernels_compute.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "defines.h"
+#include "core/defines.h"
 
 //////////////////////////////////////////////////////
 // UPDATE_ROW:
