+++ src/core/gpu_kernel/kernels_smem.h
