+++ src/core/gpu_kernel/qt_kernels.cu
